Completely fix weird read issue
There was this read issue where the computer reading the Pico would read
8 copies of the file in a single cluster. This also reduced the amount
of data each cluster could hold from 4kb to 512, which was no good.

To fix this was a long, long chase. Here's my methodology:

The computer talks to the Pico as if it were a USB. That's the whole of
pint of TinyUSB. USBs don't know their file systems, and only implement
the read/write commands necessary to work as a USB.

tusb_config.h specified that the maximum amount of bytes the computer
can get in one go is 512 bytes. This means that the computer tries to
read each cluster in 8 separate reads in different positions, which my
code was not designed for.

I did a lot of testing, and it turned out the only way to fix this was
to increase the limit to 4096 -- The size of the cluster. This was
because the computer requesting data was not designed to read data in
512 blocks, and only clung on to the last read at the very end of the
cluster.