Fully Implement Write(), Modify(), bug fixs
Fully implemented and soupled up Modify() to only write data if it is
not present, and erase if the sector isn't already erased. It also
patches the data on that sector to modify it.

Fully move FAT, Root Directory, and Data sections to flash memory. The
end on the Data section was removed to store more data.