Refactor LBA lookup for much more readable code
Read and Write commands for Fat16 are very, very simple: They just
translate the LBA address to the address where the data is stored. For
the boot section, it is read from the executabble with no write. For
everything else, it is a simple formula to translate it to flash
address.

One major consequence that doesn't effect operation is that old data
from previous runs can be read from the FAT, Root Dir, and Data sections
because they were never erased.